requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.6.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.2
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.3
distro==1.9.0
Django==4.2.1
django-bootstrap4==21.2
django-ckeditor==6.2.0
django-cleanup==8.0.0
django-cors-headers==3.7.0
django-crispy-forms==2.0
django-js-asset==2.0.0
django-rq==2.4.1
djangorestframework==3.14.0
exceptiongroup==1.2.0
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==20.0.4
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.1.5
httpcore==1.0.2
httpx==0.26.0
hyperframe==5.2.0
idna==3.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.1
langchain-community==0.0.13
langchain-core==0.1.13
langchain-openai==0.0.3
langsmith==0.0.83
libretranslatepy==2.1.1
lxml==5.1.0
marshmallow==3.20.2
# multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
openai==1.8.0
packaging==23.2
Pillow==9.5.0
polib==1.2.0
pydantic==2.5.3
pydantic_core==2.14.6
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
rq==1.15.1
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.25
sqlparse==0.4.4
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.66.1
translate==3.6.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.4
whitenoise==5.2.0
yarl==1.9.4
